requirements.txt
python-dotenv
aiogram==3.*
peewee
